Move CourtCaseListEntryHelperFunction file to the right folder
